agregue condicional para cuando la brand es undefined y corregi error de consola de logout

+++ js/logout.js
@@ -1,6 +1,8 @@
 let logout = document.querySelector("#logout");
 
-logout.addEventListener('click', function (e) {
-    localStorage.clear();   // borro el login del local storage
-    location.href = "index.html";
-})
+if (logout) {
+    logout.addEventListener('click', function (e) {
+        localStorage.clear();   // borro el login del local storage
+        location.href = "index.html";
+    })
+}
